+++ mlrun/feature_store/retrieval/online.py
@@ -16,7 +16,7 @@ import mlrun
 
 from mlrun.datastore.store_resources import ResourceCache
 from mlrun.serving.server import create_graph_server
-from mlrun.targets import get_online_target
+from mlrun.datastore.targets import get_online_target
 
 
 def _build_feature_vector_graph(
